Make add_foreign_key retriable when safe_by_default is enabled
When adding a foreign key and safe mode is enabled:
- We create a NOT VALID foreign key
- Then we commit the transaction
- Finally we validate the foreign key constraint

if the validation fails, this leaves a NOT VALID foreign key in the
database.
If we retry after having cleaned up the database, the creation of the
NOT VALID foreign key constraint will fail because it will already
be in the DB.

This commit makes the migration safely retriable by checking
if said foreign key constraint already exists, and if so,
skips its creation again.

When validation of foreign key fails, we also remove the NOT VALID
foreign key constraint as to not leave the database in an unstable
state.